HSEARCH-679 - move tests to a better package name
