+++ hibernate-search/src/test/java/org/hibernate/search/test/engine/optimizations/ItemClassBridge.java
@@ -17,7 +17,7 @@
  * MA  02110-1301, USA.
  */
 
-package org.hibernate.search.test.event;
+package org.hibernate.search.test.engine.optimizations;
 
 import org.apache.lucene.document.Document;
 import org.hibernate.search.bridge.FieldBridge;
